# Akka.Remote Performance - Write Loop iterative