Fixes #1541 

Nothing fancy, just transforming a recursive call into a iterative version.