@Aaronontheweb 2015-12-14
Perf from this test run:

# Akka.Remote.Tests.Performance.TestTransportRemoteMessagingThroughputSpec+OneWay
__Measures the throughput of Akka.Remote over a particular transport using one-way messaging__
_12/14/2015 9:10:44 PM_
### System Info
```ini
NBench=NBench, Version=0.1.5.0, Culture=neutral, PublicKeyToken=null
OS=Microsoft Windows NT 6.2.9200.0
ProcessorCount=4
CLR=4.0.30319.42000,IsMono=False,MaxGcGeneration=2
WorkerThreads=32767, IOThreads=4
```

### NBench Settings
```ini
RunMode=Iterations, TestMode=Measurement
NumberOfIterations=13, MaximumRunTime=00:00:01
```

## Data
-------------------

### Totals
          Metric |           Units |             Max |         Average |             Min |          StdDev |
---------------- |---------------- |---------------- |---------------- |---------------- |---------------- |
TotalCollections [Gen0] |     collections |            8.00 |            8.00 |            8.00 |            0.00 |
TotalCollections [Gen1] |     collections |            0.00 |            0.00 |            0.00 |            0.00 |
TotalCollections [Gen2] |     collections |            0.00 |            0.00 |            0.00 |            0.00 |
[Counter] RemoteMessageReceived |      operations |        2,436.00 |        2,407.15 |        2,365.00 |           20.40 |

### Per-second Totals
          Metric |       Units / s |         Max / s |     Average / s |         Min / s |      StdDev / s |
---------------- |---------------- |---------------- |---------------- |---------------- |---------------- |
TotalCollections [Gen0] |     collections |            7.99 |            7.99 |            7.98 |            0.00 |
TotalCollections [Gen1] |     collections |            0.00 |            0.00 |            0.00 |            0.00 |
TotalCollections [Gen2] |     collections |            0.00 |            0.00 |            0.00 |            0.00 |
[Counter] RemoteMessageReceived |      operations |        2,433.27 |        2,403.22 |        2,361.54 |           20.71 |

### Raw Data
#### TotalCollections [Gen0]
           Run # |     collections | collections / s |ns / collections |
---------------- |---------------- |---------------- |---------------- |
               1 |            8.00 |            7.99 |  125,140,325.00 |
               2 |            8.00 |            7.99 |  125,219,637.50 |
               3 |            8.00 |            7.98 |  125,290,500.00 |
               4 |            8.00 |            7.99 |  125,211,437.50 |
               5 |            8.00 |            7.98 |  125,239,862.50 |
               6 |            8.00 |            7.99 |  125,183,137.50 |
               7 |            8.00 |            7.99 |  125,156,000.00 |
               8 |            8.00 |            7.99 |  125,172,700.00 |
               9 |            8.00 |            7.99 |  125,132,937.50 |
              10 |            8.00 |            7.99 |  125,223,850.00 |
              11 |            8.00 |            7.98 |  125,245,062.50 |
              12 |            8.00 |            7.99 |  125,208,687.50 |
              13 |            8.00 |            7.98 |  125,241,037.50 |

#### TotalCollections [Gen1]
           Run # |     collections | collections / s |ns / collections |
---------------- |---------------- |---------------- |---------------- |
               1 |            0.00 |            0.00 |1,001,122,600.00 |
               2 |            0.00 |            0.00 |1,001,757,100.00 |
               3 |            0.00 |            0.00 |1,002,324,000.00 |
               4 |            0.00 |            0.00 |1,001,691,500.00 |
               5 |            0.00 |            0.00 |1,001,918,900.00 |
               6 |            0.00 |            0.00 |1,001,465,100.00 |
               7 |            0.00 |            0.00 |1,001,248,000.00 |
               8 |            0.00 |            0.00 |1,001,381,600.00 |
               9 |            0.00 |            0.00 |1,001,063,500.00 |
              10 |            0.00 |            0.00 |1,001,790,800.00 |
              11 |            0.00 |            0.00 |1,001,960,500.00 |
              12 |            0.00 |            0.00 |1,001,669,500.00 |
              13 |            0.00 |            0.00 |1,001,928,300.00 |

#### TotalCollections [Gen2]
           Run # |     collections | collections / s |ns / collections |
---------------- |---------------- |---------------- |---------------- |
               1 |            0.00 |            0.00 |1,001,122,600.00 |
               2 |            0.00 |            0.00 |1,001,757,100.00 |
               3 |            0.00 |            0.00 |1,002,324,000.00 |
               4 |            0.00 |            0.00 |1,001,691,500.00 |
               5 |            0.00 |            0.00 |1,001,918,900.00 |
               6 |            0.00 |            0.00 |1,001,465,100.00 |
               7 |            0.00 |            0.00 |1,001,248,000.00 |
               8 |            0.00 |            0.00 |1,001,381,600.00 |
               9 |            0.00 |            0.00 |1,001,063,500.00 |
              10 |            0.00 |            0.00 |1,001,790,800.00 |
              11 |            0.00 |            0.00 |1,001,960,500.00 |
              12 |            0.00 |            0.00 |1,001,669,500.00 |
              13 |            0.00 |            0.00 |1,001,928,300.00 |

#### [Counter] RemoteMessageReceived
           Run # |      operations |  operations / s | ns / operations |
---------------- |---------------- |---------------- |---------------- |
               1 |        2,436.00 |        2,433.27 |      410,969.87 |
               2 |        2,432.00 |        2,427.73 |      411,906.70 |
               3 |        2,384.00 |        2,378.47 |      420,437.92 |
               4 |        2,390.00 |        2,385.96 |      419,117.78 |
               5 |        2,392.00 |        2,387.42 |      418,862.42 |
               6 |        2,365.00 |        2,361.54 |      423,452.47 |
               7 |        2,423.00 |        2,419.98 |      413,226.58 |
               8 |        2,426.00 |        2,422.65 |      412,770.65 |
               9 |        2,405.00 |        2,402.44 |      416,242.62 |
              10 |        2,408.00 |        2,403.70 |      416,026.08 |
              11 |        2,416.00 |        2,411.27 |      414,718.75 |
              12 |        2,411.00 |        2,406.98 |      415,458.11 |
              13 |        2,405.00 |        2,400.37 |      416,602.20 |