@Aaronontheweb 2015-12-15
On my system the updated code has demonstrably lower performance - about 5,000 ns / operation.

## With change
# Akka.Remote.Tests.Performance.TestTransportRemoteMessagingThroughputSpec+OneWay
__Measures the throughput of Akka.Remote over a particular transport using one-way messaging__
_12/15/2015 6:35:31 PM_
### System Info
```ini
NBench=NBench, Version=0.1.3.0, Culture=neutral, PublicKeyToken=null
OS=Microsoft Windows NT 6.2.9200.0
ProcessorCount=8
CLR=4.0.30319.42000,IsMono=False,MaxGcGeneration=2
WorkerThreads=32767, IOThreads=8
```

### NBench Settings
```ini
RunMode=Iterations, TestMode=Measurement
NumberOfIterations=13, MaximumRunTime=00:00:01
```

## Data
-------------------

### Totals
          Metric |           Units |             Max |         Average |             Min |          StdDev |
---------------- |---------------- |---------------- |---------------- |---------------- |---------------- |
TotalCollections [Gen0] |     collections |            1.00 |            1.00 |            1.00 |            0.00 |
TotalCollections [Gen1] |     collections |            0.00 |            0.00 |            0.00 |            0.00 |
TotalCollections [Gen2] |     collections |            0.00 |            0.00 |            0.00 |            0.00 |
[Counter] RemoteMessageReceived |      operations |        3,750.00 |        3,699.77 |        3,541.00 |           53.49 |

### Per-second Totals
          Metric |       Units / s |         Max / s |     Average / s |         Min / s |      StdDev / s |
---------------- |---------------- |---------------- |---------------- |---------------- |---------------- |
TotalCollections [Gen0] |     collections |            1.02 |            1.00 |            1.00 |            0.01 |
TotalCollections [Gen1] |     collections |            0.00 |            0.00 |            0.00 |            0.00 |
TotalCollections [Gen2] |     collections |            0.00 |            0.00 |            0.00 |            0.00 |
[Counter] RemoteMessageReceived |      operations |        3,744.86 |        3,704.56 |        3,593.77 |           38.56 |

### Raw Data
#### TotalCollections [Gen0]
           Run # |     collections | collections / s |ns / collections |
---------------- |---------------- |---------------- |---------------- |
               1 |            1.00 |            1.00 |1,001,106,719.77 |
               2 |            1.00 |            1.00 |1,001,371,425.93 |
               3 |            1.00 |            1.02 |  985,164,053.67 |
               4 |            1.00 |            1.00 |1,000,960,040.92 |
               5 |            1.00 |            1.00 |1,001,069,515.52 |
               6 |            1.00 |            1.00 |1,001,407,347.29 |
               7 |            1.00 |            1.01 |  985,317,147.06 |
               8 |            1.00 |            1.00 |1,000,838,164.90 |
               9 |            1.00 |            1.00 |1,001,418,465.80 |
              10 |            1.00 |            1.00 |1,000,711,584.89 |
              11 |            1.00 |            1.00 |1,000,706,453.27 |
              12 |            1.00 |            1.00 |1,001,375,702.28 |
              13 |            1.00 |            1.00 |1,001,347,050.73 |

#### TotalCollections [Gen1]
           Run # |     collections | collections / s |ns / collections |
---------------- |---------------- |---------------- |---------------- |
               1 |            0.00 |            0.00 |1,001,106,719.77 |
               2 |            0.00 |            0.00 |1,001,371,425.93 |
               3 |            0.00 |            0.00 |  985,164,053.67 |
               4 |            0.00 |            0.00 |1,000,960,040.92 |
               5 |            0.00 |            0.00 |1,001,069,515.52 |
               6 |            0.00 |            0.00 |1,001,407,347.29 |
               7 |            0.00 |            0.00 |  985,317,147.06 |
               8 |            0.00 |            0.00 |1,000,838,164.90 |
               9 |            0.00 |            0.00 |1,001,418,465.80 |
              10 |            0.00 |            0.00 |1,000,711,584.89 |
              11 |            0.00 |            0.00 |1,000,706,453.27 |
              12 |            0.00 |            0.00 |1,001,375,702.28 |
              13 |            0.00 |            0.00 |1,001,347,050.73 |

#### TotalCollections [Gen2]
           Run # |     collections | collections / s |ns / collections |
---------------- |---------------- |---------------- |---------------- |
               1 |            0.00 |            0.00 |1,001,106,719.77 |
               2 |            0.00 |            0.00 |1,001,371,425.93 |
               3 |            0.00 |            0.00 |  985,164,053.67 |
               4 |            0.00 |            0.00 |1,000,960,040.92 |
               5 |            0.00 |            0.00 |1,001,069,515.52 |
               6 |            0.00 |            0.00 |1,001,407,347.29 |
               7 |            0.00 |            0.00 |  985,317,147.06 |
               8 |            0.00 |            0.00 |1,000,838,164.90 |
               9 |            0.00 |            0.00 |1,001,418,465.80 |
              10 |            0.00 |            0.00 |1,000,711,584.89 |
              11 |            0.00 |            0.00 |1,000,706,453.27 |
              12 |            0.00 |            0.00 |1,001,375,702.28 |
              13 |            0.00 |            0.00 |1,001,347,050.73 |

#### [Counter] RemoteMessageReceived
           Run # |      operations |  operations / s | ns / operations |
---------------- |---------------- |---------------- |---------------- |
               1 |        3,746.00 |        3,741.86 |      267,246.86 |
               2 |        3,750.00 |        3,744.86 |      267,032.38 |
               3 |        3,666.00 |        3,721.21 |      268,729.97 |
               4 |        3,734.00 |        3,730.42 |      268,066.43 |
               5 |        3,713.00 |        3,709.03 |      269,612.04 |
               6 |        3,737.00 |        3,731.75 |      267,970.93 |
               7 |        3,541.00 |        3,593.77 |      278,259.57 |
               8 |        3,697.00 |        3,693.90 |      270,716.30 |
               9 |        3,690.00 |        3,684.77 |      271,387.12 |
              10 |        3,694.00 |        3,691.37 |      270,901.89 |
              11 |        3,709.00 |        3,706.38 |      269,804.92 |
              12 |        3,719.00 |        3,713.89 |      269,259.40 |
              13 |        3,701.00 |        3,696.02 |      270,561.21 |




## Without change
----
# Akka.Remote.Tests.Performance.TestTransportRemoteMessagingThroughputSpec+OneWay
__Measures the throughput of Akka.Remote over a particular transport using one-way messaging__
_12/15/2015 6:11:02 PM_
### System Info
```ini
NBench=NBench, Version=0.1.3.0, Culture=neutral, PublicKeyToken=null
OS=Microsoft Windows NT 6.2.9200.0
ProcessorCount=8
CLR=4.0.30319.42000,IsMono=False,MaxGcGeneration=2
WorkerThreads=32767, IOThreads=8
```

### NBench Settings
```ini
RunMode=Iterations, TestMode=Measurement
NumberOfIterations=13, MaximumRunTime=00:00:01
```

## Data
-------------------

### Totals
          Metric |           Units |             Max |         Average |             Min |          StdDev |
---------------- |---------------- |---------------- |---------------- |---------------- |---------------- |
TotalCollections [Gen0] |     collections |            1.00 |            1.00 |            1.00 |            0.00 |
TotalCollections [Gen1] |     collections |            0.00 |            0.00 |            0.00 |            0.00 |
TotalCollections [Gen2] |     collections |            0.00 |            0.00 |            0.00 |            0.00 |
[Counter] RemoteMessageReceived |      operations |        3,776.00 |        3,753.15 |        3,654.00 |           32.41 |

### Per-second Totals
          Metric |       Units / s |         Max / s |     Average / s |         Min / s |      StdDev / s |
---------------- |---------------- |---------------- |---------------- |---------------- |---------------- |
TotalCollections [Gen0] |     collections |            1.00 |            1.00 |            0.99 |            0.00 |
TotalCollections [Gen1] |     collections |            0.00 |            0.00 |            0.00 |            0.00 |
TotalCollections [Gen2] |     collections |            0.00 |            0.00 |            0.00 |            0.00 |
[Counter] RemoteMessageReceived |      operations |        3,771.49 |        3,747.69 |        3,635.49 |           36.09 |

### Raw Data
#### TotalCollections [Gen0]
           Run # |     collections | collections / s |ns / collections |
---------------- |---------------- |---------------- |---------------- |
               1 |            1.00 |            1.00 |1,001,084,910.38 |
               2 |            1.00 |            0.99 |1,005,092,279.39 |
               3 |            1.00 |            1.00 |1,000,759,052.39 |
               4 |            1.00 |            1.00 |1,000,974,580.51 |
               5 |            1.00 |            1.00 |1,001,275,208.02 |
               6 |            1.00 |            1.00 |1,001,190,963.90 |
               7 |            1.00 |            1.00 |1,001,339,780.93 |
               8 |            1.00 |            1.00 |1,000,931,389.36 |
               9 |            1.00 |            1.00 |1,001,134,516.06 |
              10 |            1.00 |            1.00 |1,001,099,877.61 |
              11 |            1.00 |            1.00 |1,001,215,339.10 |
              12 |            1.00 |            1.00 |1,001,450,110.80 |
              13 |            1.00 |            1.00 |1,001,535,210.20 |

#### TotalCollections [Gen1]
           Run # |     collections | collections / s |ns / collections |
---------------- |---------------- |---------------- |---------------- |
               1 |            0.00 |            0.00 |1,001,084,910.38 |
               2 |            0.00 |            0.00 |1,005,092,279.39 |
               3 |            0.00 |            0.00 |1,000,759,052.39 |
               4 |            0.00 |            0.00 |1,000,974,580.51 |
               5 |            0.00 |            0.00 |1,001,275,208.02 |
               6 |            0.00 |            0.00 |1,001,190,963.90 |
               7 |            0.00 |            0.00 |1,001,339,780.93 |
               8 |            0.00 |            0.00 |1,000,931,389.36 |
               9 |            0.00 |            0.00 |1,001,134,516.06 |
              10 |            0.00 |            0.00 |1,001,099,877.61 |
              11 |            0.00 |            0.00 |1,001,215,339.10 |
              12 |            0.00 |            0.00 |1,001,450,110.80 |
              13 |            0.00 |            0.00 |1,001,535,210.20 |

#### TotalCollections [Gen2]
           Run # |     collections | collections / s |ns / collections |
---------------- |---------------- |---------------- |---------------- |
               1 |            0.00 |            0.00 |1,001,084,910.38 |
               2 |            0.00 |            0.00 |1,005,092,279.39 |
               3 |            0.00 |            0.00 |1,000,759,052.39 |
               4 |            0.00 |            0.00 |1,000,974,580.51 |
               5 |            0.00 |            0.00 |1,001,275,208.02 |
               6 |            0.00 |            0.00 |1,001,190,963.90 |
               7 |            0.00 |            0.00 |1,001,339,780.93 |
               8 |            0.00 |            0.00 |1,000,931,389.36 |
               9 |            0.00 |            0.00 |1,001,134,516.06 |
              10 |            0.00 |            0.00 |1,001,099,877.61 |
              11 |            0.00 |            0.00 |1,001,215,339.10 |
              12 |            0.00 |            0.00 |1,001,450,110.80 |
              13 |            0.00 |            0.00 |1,001,535,210.20 |

#### [Counter] RemoteMessageReceived
           Run # |      operations |  operations / s | ns / operations |
---------------- |---------------- |---------------- |---------------- |
               1 |        3,750.00 |        3,745.94 |      266,955.98 |
               2 |        3,654.00 |        3,635.49 |      275,066.31 |
               3 |        3,766.00 |        3,763.14 |      265,735.28 |
               4 |        3,757.00 |        3,753.34 |      266,429.22 |
               5 |        3,730.00 |        3,725.25 |      268,438.39 |
               6 |        3,775.00 |        3,770.51 |      265,216.15 |
               7 |        3,776.00 |        3,770.95 |      265,185.32 |
               8 |        3,775.00 |        3,771.49 |      265,147.39 |
               9 |        3,764.00 |        3,759.73 |      265,976.23 |
              10 |        3,758.00 |        3,753.87 |      266,391.67 |
              11 |        3,765.00 |        3,760.43 |      265,927.05 |
              12 |        3,770.00 |        3,764.54 |      265,636.63 |
              13 |        3,751.00 |        3,745.25 |      267,004.85 |